Break out A3 and A4
